+++ packages/puppeteer-core/Herebyfile.mjs
@@ -14,6 +14,30 @@ import {task} from 'hereby';
 
 const require = Module.createRequire(import.meta.url);
 
+export const updateVersionTask = task({
+  name: 'update:version',
+  run: async () => {
+    // x-release-please-start-version
+    const version = '24.34.0';
+    // x-release-please-end
+
+    // We only want to do this once we are trying to publish
+    // a new version
+    if (!process.env['PUBLISH']) {
+      return;
+    }
+
+    const fileContent = await readFile('../../versions.json', {
+      encoding: 'utf-8',
+    });
+
+    await writeFile(
+      '../../versions.json',
+      fileContent.replace(`"NEXT"`, `"v${version}"`),
+    );
+  },
+});
+
 export const generateInjectedTask = task({
   name: 'generate:injected',
   run: async () => {
@@ -47,7 +71,11 @@ export const generatePackageJsonTask = task({
 
 export const generateTask = task({
   name: 'generate',
-  dependencies: [generateInjectedTask, generatePackageJsonTask],
+  dependencies: [
+    generateInjectedTask,
+    generatePackageJsonTask,
+    updateVersionTask,
+  ],
 });
 
 export const buildTscTask = task({