+++ packages/puppeteer-core/src/api/Page.ts
@@ -32,7 +32,6 @@ import type {HTTPRequest} from '../api/HTTPRequest.js';
 import type {HTTPResponse} from '../api/HTTPResponse.js';
 import type {Accessibility} from '../cdp/Accessibility.js';
 import type {Coverage} from '../cdp/Coverage.js';
-import type {DeviceRequestPrompt} from '../cdp/DeviceRequestPrompt.js';
 import type {NetworkConditions} from '../cdp/NetworkManager.js';
 import type {Tracing} from '../cdp/Tracing.js';
 import type {ConsoleMessage} from '../common/ConsoleMessage.js';
@@ -82,9 +81,11 @@ import {
 } from '../util/disposable.js';
 import {stringToTypedArray} from '../util/encoding.js';
 
-import type {Browser} from './Browser.js';
+import type {BluetoothEmulation} from './BluetoothEmulation.js';
+import type {Browser, WindowId} from './Browser.js';
 import type {BrowserContext} from './BrowserContext.js';
 import type {CDPSession} from './CDPSession.js';
+import type {DeviceRequestPrompt} from './DeviceRequestPrompt.js';
 import type {Dialog} from './Dialog.js';
 import type {
   BoundingBox,
@@ -306,7 +307,7 @@ export interface ScreenshotOptions {
    * relative to current working directory. If no path is provided, the image
    * won't be saved to the disk.
    */
-  path?: `${string}.${ImageFormat}`;
+  path?: string;
   /**
    * Specifies the region of the page/element to clip.
    */
@@ -1455,10 +1456,8 @@ export abstract class Page extends EventEmitter<PageEvents> {
   async $$eval<
     Selector extends string,
     Params extends unknown[],
-    Func extends EvaluateFuncWith<
-      Array<NodeFor<Selector>>,
-      Params
-    > = EvaluateFuncWith<Array<NodeFor<Selector>>, Params>,
+    Func extends EvaluateFuncWith<Array<NodeFor<Selector>>, Params> =
+      EvaluateFuncWith<Array<NodeFor<Selector>>, Params>,
   >(
     selector: Selector,
     pageFunction: Func | string,
@@ -2705,6 +2704,13 @@ export abstract class Page extends EventEmitter<PageEvents> {
   }
 
   /**
+   * Emulates focus state of the page.
+   *
+   * @param enabled - Whether to emulate focus.
+   */
+  abstract emulateFocusedPage(enabled: boolean): Promise<void>;
+
+  /**
    * @internal
    */
   abstract _screenshot(options: Readonly<ScreenshotOptions>): Promise<string>;
@@ -3148,16 +3154,22 @@ export abstract class Page extends EventEmitter<PageEvents> {
   ): Promise<DeviceRequestPrompt>;
 
   /**
-   * Resizes the browser window the page is in so that the content area
-   * (excluding browser UI) is according to the specified widht and height.
+   * Resizes the browser window of this page so that the content area (excluding
+   * browser UI) has the specified width and height.
    *
    * @experimental
-   * @internal
    */
   abstract resize(params: {
     contentWidth: number;
     contentHeight: number;
   }): Promise<void>;
+
+  /**
+   * Returns the page's window id.
+   *
+   * @experimental
+   */
+  abstract windowId(): Promise<WindowId>;
 
   /** @internal */
   override [disposeSymbol](): void {
@@ -3168,6 +3180,17 @@ export abstract class Page extends EventEmitter<PageEvents> {
   [asyncDisposeSymbol](): Promise<void> {
     return this.close();
   }
+
+  /**
+   * Opens DevTools for the current Page and returns the DevTools Page. This
+   * method is only available in Chrome.
+   */
+  abstract openDevTools(): Promise<Page>;
+
+  /**
+   * {@inheritDoc BluetoothEmulation}
+   */
+  abstract get bluetooth(): BluetoothEmulation;
 }
 
 /**